+++ src/App.css
@@ -32,14 +32,15 @@
   }
 }
 .table{
-  border:solid 1px gray;
-  width: 500px;
-  height: 500px;
-  table-layout:center;
-  margin-left:500px;
-  margin-top:50px;
+  padding-top: 10px;
+  border:solid 0px gray;
+  width: 50%;
+  height: 450px;
+  margin-left: 25%;
+  margin-top:3%;
   border-radius: 20px;
-  
+  background-color: white;
+  box-shadow: 2px 2px 6px 6px #D3D3D3;   
 }
 .paper{
   margin-Top: 50px ;
@@ -52,7 +53,7 @@
  color: red;
 }
 .form{
-  width: 400px;
+  width: 100%;
    margin-Top: 10px;
 }
 .submit{
@@ -66,10 +67,13 @@
   background-color: aquamarine;
 }
 .paper1 {
+  border: solid 0px red;
   padding:20px;
   text-Align: center;
- 
-
+  background-color: #F5F5F5;
+  
+  
+  
 }
 .root {
   flex-Grow: 1;
