+++ src/routes.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom'
-import HomePageContainer from './containers/HomePage.container';
+import HomePage from './containers/HomePage';
 
 export default (
   <Router>
-    <Route path='/' component={HomePageContainer} />
+    <Route path='/' component={HomePage} />
   </Router>
 
 );
